requirements.txt

Flask==2.3.2
flask-socketio==5.3.5
openai==0.27.8
python-dotenv==1.0.0